Clean up dependency organization
Dependencies are organized into three major categories:
1) Project dependencies
2) External dependencies
3) Test scope dependencies

(A fourth category is "Runtime scope dependencies" but nothing in
ImgLib2 needs that scope.)

Furthermore, dependencies that are not available (or planned to be
available soon) from Central are flagged with FIXME, so that do not
forget to resolve them somehow.

These dependencies can be seen with:

    grep -A 5 -B 1 FIXME $(git ls-files '*pom.xml')

And further grepping for artifactId reveals the following artifact
dependencies which we will need to somehow address:

    loci-common, mines-jtk, mpicbg, udunits

Fortunately, none of these is a dependency for the most critical core
ImgLib2 components.