Update to the ImgLib2 stable release
This uses the pom-imglib2 parent to inherit common configuration blocks.
We no longer need to declare most versions, since they are managed.

This also renames the artifact to imglib2-script, for consistency with
its package prefix, and restarts the versioning at 0.1.0-SNAPSHOT.

It also makes the necessary code changes to adjust for the backwards
incompatibilities of the ImgLib2 release.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>